Adds drop location spoilers to the Spookyraven Wine Cellar, along the lines of my previous Barrel Full Of Barrels code.  The "Wine Cellar (Random)" adv location has been renamed as "Wine Cellar (automatic)", as it now chooses a corner based on any dusty wines in your conditions.
There is one big difference between the Cellar and Barrels - nobody has really spaded the possible
patterns of the Cellar wines, beyond the fact that there are only four possible sets of wine that can
drop in any one corner.  This is due to a combination of the fact that it's a once-per-ascension piece
of data, and some game bugs that made it uncertain as to which corner you were actually adventuring in.
Therefore, for now the code assumes that every possible drop pattern is equally likely, but it remembers
which pattern it found, and can kmail that information to me if you type "spade" in the CLI.

The spade mechanism is designed to be extensible and privacy-preserving.  It can queue up multiple items
of data, from the same or different sources, and the user will always be shown the full data, its
recipient, and an explanation of what the data will be used for, before anything is transmitted.  A
nudge will be given at login if there's data waiting to be sent, but it can be ignored without
consequence.

Note for testers: please adventure in the Wine Cellars (assuming that you can get to them) until the
spoiler text indicates only three possible wines in each location.  Two distinct wine drops, in any three
of the corners, will be suffient.  Then type "spade", read the info, and I hope you'll be willing to
say "yes" to it.  Also in need of testing: when you first visit the pedestal in the Cellar with Lord
Spookyraven's spectacles equipped, the new code should set "Wine Cellar (automatic)" as your adventure
location, and add the three needed dusty wines to your conditions.  I can't test that as I've already
opened the Summoning Chamber this ascension.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@6360 b29ace70-8910-0410-8dcc-aa2fc6433167