Properly summons a s'n's sword when the first sticker is used.  Note that this sticker will end up in slot 1, no matter where you were trying to put it - that's just the way the summoning works.
Makes the estimated number of sticker turns remaining visible in the Gear
Changer.

Adds a "stickers <st1>, <st2>, <st3>" command to the CLI, which applies
stickers in currently empty slots only.  This is intended for use as the
command of an unconditional mood trigger; for example, "stickers wrestler,
wrestler, wrestler" will keep your stats maximized as stickers wear out.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@6600 b29ace70-8910-0410-8dcc-aa2fc6433167