Extends the recent locations history in the relay browser - by default, 5 locations (settable via recentLocations) are remembered, and links shown underneath the existing last adventure link in smaller text.
Various improvements to the Daily Deeds panel:

Generalized the preference change notification mechanism so that deeds can
update in response to changes that aren't actually stored in a preference -
for example, a change in the number of stills remaining is broadcast as a
change to the nonexistent preference "(stills)".

Deeds are now executed via the command queue, so it's safe to click multple
deeds without waiting.

User-entered deeds are now dimmed out in both Daily Deeds lists upon
successful completion.

Added Deeds for remaining stills, outrageous sombrero, and black puddings. 


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@6625 b29ace70-8910-0410-8dcc-aa2fc6433167