Spooky Putty Items are "used" via the Item Manager, not equipped.
Add initial support for The Marinara Trench. New items, concoctions, skills,
etc. etc.

I've only seen the Sauceress choice adventure so far, so the Pastamancer one is
not yet supported.

Most of the new status effects are not yet on the Wiki, so they, too, are not
present.

For that matter, I have no consumption data for Tempura, yet, although it
should be makable.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@6751 b29ace70-8910-0410-8dcc-aa2fc6433167