On the page where you can jump through the Gash, remind player to make quest keys into key lime pies and to use their bubblin' stone, if necessary.
Add the preferredWebBrowser preference to the Browser tab of the Preferences.
Provide special help text documenting it for OS X, Linux, and Windows.
For the latter two OS's, provide a button which pops up a Choose File dialog
to allow you to browse your file system to find your browser of choice.

Since I don't have either Linux or Windows, that is untested.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@6780 b29ace70-8910-0410-8dcc-aa2fc6433167