The "csend" command can now be used to sent meat to anybody, not just known buffbots.  As always, it won't send to buffbots that have requested to be excluded (currently, only Noblesse Oblige is in this category), and will upgrade any request for a philanthropic buff to the nearest non-philanthropic buff amount.
Added 3 ways to specify a price limit when buying items:
1. CLI buy command can now be given as "buy [<num>] <item> @ <limit>" - the
limit is the maximum price that will be paid for any one item.

2. New form of ASH function: int buy(int qty, item itm, int limit)
Does the same as the CLI command, with limit=0 indicating no price limit
(equivalent to the 2-arg version of buy).  The return value is the number of
items purchased.

3. The price limit for implicitly purchased items (such as missing
ingredients for recipes), formerly hard-coded as 20,000, is now settable via
preference "autoBuyPriceLimit".

Now, let's get rid of all that "closet all but X meat" crap...


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@6844 b29ace70-8910-0410-8dcc-aa2fc6433167