Adds ASH mall_price(item) function, which returns the approximate price at which the item can be bought.  To prevent abuse, this will only perform an actual Mall search once per item per session; subsequent calls for the same item will return a cached value.  The cache will be updated after a successful purchase of the item, but not in any other conditions. Furthermore, the price returned will be that of the 5th one available, which should be representative of the actual cost of buying a reasonable quantity - scripts will not be able to take advantage of (or be confused by) mispriced items.
Mall searches for a specific item are now cached, with a 15 second lifespan. 
When comparison shopping, this will allow the chosen item to be purchased
without having to perform another search for it.

Added new CLI commands "cheapest" and "expensive", which take a
comma-separated list of items.  On a line by themselves, they will look up
pricing for all the items (subject to the caching described above), and
display a sorted list.  If there are further commands on the same line, those
commands will be executed with every occurence of the word "it" being
replaced by the name of the best item (cheapest or most expensive, depending
on which version of the command you used).  Example:
     "cheapest hi mein; buy 3 it; eat 3 it"
For convenience, these commands accept multiple matches (so the "hi mein" in
the example actually matches all five flavors, rather than producing an
ambiguous name error), and will accept items preceeded by a minus sign to
exclude them from the match (so "cheapest hi mein, -sleazy" would match the
other four flavors only).


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@6989 b29ace70-8910-0410-8dcc-aa2fc6433167