With a non-zero pull budget set, allow Item Manager consumption lists to show suitable items from storage even if they are not otherwise creatable (the conditions were being checked in the wrong order).
Add [council] links for Boss Bat Bandana, Skull of the Bonerdagon.

In the Modifier Maximizer, don't show grayed-out consumables that you don't
have room for, if you don't actually have the items (another instance of
checking conditions in the wrong order).

Add Emblem of Ak'gyxoth to the list of MP cost reducing items (mafia was
replacing it with a navel ring, which is also -1 MP).


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@7693 b29ace70-8910-0410-8dcc-aa2fc6433167