When displaying counters in the relay browser, handle any unexpected formatting of the charpane without crashing.
Add usability of the moveable feast (although we're still not doing any
tracking whatsoever of the resulting buff).

Add modifiers for the Floaty Fatigues and Vestments of the Treeslayer.

Add the synergies between the six 'Secret from the Future' items.  This is
partially speculative, as nobody seems to be actively spading these any more.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@7788 b29ace70-8910-0410-8dcc-aa2fc6433167