Add experimental support for the Great Big Chessboard: - When you first see the chessboard, KoLmafia saves the configuration. - It adds a "Solve!" button to the page. For the moment, that does nothing except calls "chess solve" which brings you back to the reflection of a map. - Having seen the map, in the gCLI you can use the "chess" command to do things - "chess board" - shows you the board - "chess test" - solves the puzzle and prints out the sequence of moves - "chess solve" - does nothing, yet, but will solve and execute the solution.
As coded, a solution requires that you be able to move the last piece standing
all the way to the final row in a single move, after clearing all the other
pieces off the board. I don't know if the puzzle actually insists on that or
if there might be versions in which a piece other than a queen or rook might 
have to take multiple moves to finish.

If you use "chess test" and are told that it couldn't find a solution, please
walk away from the chessboard, turn on debug logging, go there again, and send
me the log. Thanks.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@8236 b29ace70-8910-0410-8dcc-aa2fc6433167