It's March Modifier Madness time!
* Added mainhand(text) function to the modifier expression language, to check
your weapon type.  Used this to add the newly-discovered bonus to Expert
Panhandling when wielding a saucepan.

* Added "Candy Drop" numeric modifier, because hey - free candy.

* Added "Familiar Experience" and "DB Combat Damage" numeric modifiers.

* Added "Sombrero Bonus" modifier, to handle the unique "+10 lbs. of
Sombrero" given by the mariachihuahua's familiar equipment.  (New modifiers
for "+10 lbs. of Volley/Lep/Fairy" aren't needed, since those are expressible
in terms of existing modifiers.)

* Added "Sporadic Meat Drop" and "Sporadic Item Drop", to specify the average
amounts of bonuses that don't fire in every combat.  These are NOT included
in your basic meat/item modifiers (since that would screw up some
calculations, such as determining whether an item drop is guaranteed), but
maximizing those modifiers will include these values in the score.

* Added "Meat Bonus" for sources of a flat meat drop increase, instead of the
usual percentage increase - currently just the ant tools.  For now, the
maximizer considers these to be equivalent to a tiny amount of +Meat%, so
that the ant tools can be chosen in preference to familiar equipment that
doesn't increase meat drop at all.  A more accurate comparison could be done
by considering the average meat drop in the current location.

* If the meatDropSpading feature triggers when you have a nonzero Sporadic
Meat Drop or Meat Bonus, this is indicated in the results, since this is
almost certainly the source of the discrepancy.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@8332 b29ace70-8910-0410-8dcc-aa2fc6433167