Avoid problems with getting the jump with a combat macro autoattack set.
Added temporary debugging feature: "set macroLens=true" will reveal a text
field in the combat form (non-CAB only) in which you can directly type a
combat macro to execute.  You wouldn't want to actually play the game this
way, but it's a far more convenient way of experimenting with macros than
constantly going to the macro editor.  You must have at least one saved macro
for this to work, and mafia's combat hotkeys are disabled - you wouldn't be
able to type numbers in the field, otherwise.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@8397 b29ace70-8910-0410-8dcc-aa2fc6433167