Add two parameter version of maximize CLI command and four parameter version of the internal maximize routine as ASH functions.
In the two parameter version, the first parameter is the maximizer string and the second parameter is true if you need the equivalent of the maximize? command.

In the four parameter version, there are two integers that have something to do with price.  However, I don't know anything about modifier maximizer other than its return values, so you'll have to poke someone else to explain it to you.  Or you can just pass zero for both integer values, which is what the two parameter version does.

git-svn-id: https://svn.code.sf.net/p/kolmafia/code@8954 b29ace70-8910-0410-8dcc-aa2fc6433167