Accessories and off-hand items have a power rating, just like other equipment, but it is not visible in the item description. However, api.php will reveal it to you - if you happen to have the item in inventory or your closet.
When we look up an unknown item from one of those sources, fetch the power and
include it in the data override.

Add a "checkpowers" command which will look at all accessories and off-hand
items in closet and inventory and, if we have the power listed as 0, attempt
to register them with an override, which will look up the item in api.php.

I ran this on my main and found a bunch of accessory and offhand item powers.
Update equipment.txt with that data. Much more remains to be discovered.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@9080 b29ace70-8910-0410-8dcc-aa2fc6433167