If you have autoSatisfyWithNPCs set, assign a "cost" to worthless items. If you have all 13 starter items available, the cost is fixed: 50 Meat for a chewing gum on a string. If you have fewer than 13, the number of pieces of chewing gum you will need to use is probabilistic. Define the cost as 50 Meat times the "expected" number of pieces of gum needed - which is as high as 4.25, if you have no sewer items yet.
Concoctions using items available from the Hermit are now creatable even if you
curently have no worthless items: the "ingredient" of a jaba pepper, say, is a
worthless item, and "acquire" will go fishing for a worthless item, if neded,
for the expected cost as described above.

Note that this is not QUITE enough; if you haven't hax0red the Hermit, a Hermit
permit is also an implicit ingredient for any hermit item. We don't handle that
yet. Coming soon...


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@9690 b29ace70-8910-0410-8dcc-aa2fc6433167