Turns out that interrupt() doesn't do what I'd hope it would do, so don't bother calling out.
However, there is a better way to print out what's actually going on inside of KoLmafia at the time the 'graygui' command is executed by printing the output of jstack (which is only available in Java 1.5 or higher).  Let's go ahead and do that for now, as it will give us better insight into what's going wrong.

git-svn-id: https://svn.code.sf.net/p/kolmafia/code@10494 b29ace70-8910-0410-8dcc-aa2fc6433167