Add the remaining cases where a function can be considered to properly return without an explicit "return" at the top level: "switch" with a default case and no breaks, "try" where either the main code block or "finally" clause can't fall through, and infinite loops (while(true), repeat...until(false)) with no breaks.  Note that it turned out not to be appropriate to consider abort() to be a proper return: I'd forgotten about the possibility of doing "disable abort" from the CLI.
The determination of whether a non-void function has a missing return value
should be entirely correct now, however it just gives a warning for boolean
functions at the moment, to give people a chance to fix invalid scripts that
were working only because of the previous broken determination.

The infrastructure for detecting proper returns also allows detecting
unreachable code, which generates a warning (not an error, I believe Java is
too draconian in this respect).  Unreachable returns are allowed without a
warning for now, since they were previously required to get functions to
compile if they didn't return from the top level. 


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@10727 b29ace70-8910-0410-8dcc-aa2fc6433167