Add _mediumSiphons preference.
New $familiar proxy record: .charges.  The only familiar that uses this right now is the Happy Medium, but the stomping boots and grinder can eventually utilize it.

When we look at the charpane, set .charges when we see the happy medium's image.  Unfortunately, api.php is bugged at the moment and always returns the un-charged image name.  I put the code in to do the updating, but left it commented out.

git-svn-id: https://svn.code.sf.net/p/kolmafia/code@10728 b29ace70-8910-0410-8dcc-aa2fc6433167