When you acquire a folder, add it as potential equipment to all five folder slots.
If you cannot eat, do not consider Spaghetti Breakfast as a breakfast summon.
If you cannot drink, do not consider Grab a Cold One as a breakfast summon
KOLHS is considered to be a "can't drink" path for the purpose of breakfast.

Add plurals to a bunch of KOLHS items.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@12907 b29ace70-8910-0410-8dcc-aa2fc6433167