Since The Smile of Mr. A. no longer appears on the Char Sheet, if the player has any Golden Mr. Accessories, add the skill to available skills. Also do that when the player acquires her first Golden Mr. Accessory.
The Temorary Lycanthropy effect (from Blood of the Wereseal) gives you a non-
integral percentage boost to Muscle. KoL truncates that value when generating
the effect Description, but actually uses the full value when adjusting your
Muscle. Note that it does NOT round the percentage value before applying it,
contrary to what the Wiki says and what KoLmafia previously did.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@14456 b29ace70-8910-0410-8dcc-aa2fc6433167