Support crystal ball predictions
Currently any existing prediction is invalidated when a new encounter is... encountered. Then if the end of that encounter results in a new prediction, we log it. If spading lets us hone this then the predictions will probably just get better while the API for users of scripters will remain the same.

git-svn-id: https://svn.code.sf.net/p/kolmafia/code@20596 b29ace70-8910-0410-8dcc-aa2fc6433167