Reverse Yoda-style comparisons in Parser.java.
https://en.wikipedia.org/wiki/Yoda_conditions

Courtesy of fredg1 at kolmafia.us, as part of a broader refactor as
part of adding language server features for ASH.

In particular, this is useful in Java for null-safe comparisons, at
the expense of code that reads more awkwardly. (It's often widely used
to avoid accidentally performing an assignment instead of an equality
check.)

we've looked at the logic for Parser.currentToken() and concluded that
it can never return null; thus, this should be a safe change to make.

As an added bonus, fredg also fixed a potential NullPointerException
in Parser.parseJavaFor in making this change across the board.

This will be useful in simplifying a follow-up change that refactors
Parser to use Line and Token classes.

git-svn-id: https://svn.code.sf.net/p/kolmafia/code@20815 b29ace70-8910-0410-8dcc-aa2fc6433167