Add a rudimentary ParserTest.
This tests whether a provided script is well-formed, and if not, what
error message is raised.

This also disables a couple of very slow ConcoctionTest cases, and
upgrades JUnit to 4.13 for access to assertThrows.

(I have no objections to upgrading to JUnit 5, but my impression is
that we'll need to rewrite our tests for that. If we want to do it, we
should do it sooner rather than later.)

git-svn-id: https://svn.code.sf.net/p/kolmafia/code@20816 b29ace70-8910-0410-8dcc-aa2fc6433167