Fix coverage issues with last commit.
The previous commit seems like it resulted in some backsliding in
behavior, namely we lost instrumentation of code files if we did a
clean recompilation of the source files.

I've checked multiple times that this continues to instrument the
files in question from a clean build as well as from an incremental
build.

git-svn-id: https://svn.code.sf.net/p/kolmafia/code@20822 b29ace70-8910-0410-8dcc-aa2fc6433167