Migrate new Float(float) etc to valueOf().
valueOf has been available since Java 1.5, and new Float(float) and
friends have been deprecated since Java 9.

Per Oracle docs:

"It is rarely appropriate to use this constructor. The static factory
valueOf(float) is generally a better choice, as it is likely to yield
significantly better space and time performance."

(In general, valueOf performs caching as appropriate, thereby reducing
or eliminating the need for constructions like IntegerPool.)

(This is commit 2 of 2 of "fix -Xlint:deprecation" warnings.)

git-svn-id: https://svn.code.sf.net/p/kolmafia/code@20833 b29ace70-8910-0410-8dcc-aa2fc6433167