Disallow comments broken by newlines.
Because of a quirk of how plural typed constants are parsed, the
following two snippets of code were equivalent:

$booleans[tr/
/ue];

$booleans[tr//ue];

and would yield a parse error for unexpected end-of-file.

Now, it still yields a parse error, except now it's that "tr//ue"
isn't a real boolean. Which is maybe less confusing?

Either way, courtesy of fredg1 (as was the last commit).

git-svn-id: https://svn.code.sf.net/p/kolmafia/code@20837 b29ace70-8910-0410-8dcc-aa2fc6433167