Rewrite various conditionals in Parser.
Courtesy of fredg1, as yet another step on the road to language server
support for ASH -- the code is slowly moving toward a state where the
Parser will be able to collect all the parse errors in a file, as
opposed to terminating after the first one.

In particular, this patch refactors a number of instances of

if ( foo ) throw bar;
baz;

into

if ( !foo )
{
  baz;
}
else
{
  throw bar;
}

(There are some other minor changes, but per test coverage, I'm
reasonably sure that behavior hasn't changed meaningfully, other than
one error message changing for the better.)

This change also adds a _lot_ of test cases to cover the affected
branches (and in some cases, I went ahead and finished covering the
entire method). Parser.java coverage is now above 80%. Terrifying.

git-svn-id: https://svn.code.sf.net/p/kolmafia/code@20883 b29ace70-8910-0410-8dcc-aa2fc6433167